WIP: First attempt to get xpack build working
The build is still non-functional due to undefined symbols at the linker stage.
I suspect this is due to differences in the xpack gcc vs the package downloaded from Arm.com.
It may be necessary to resolve the --specs=nano.specs and --specs=nosys.specs issues for this to work.
Additionally, some of the setup lines were removed in the arm-none-eabi-gcc toolchain file since it seems to actually prevent proper identification of binutils loading into CMakeCache.txt.