Got xpack dev environment management working for debug and release builds
The package.json was updated to explicitly specify both Debug and Release configurations.
Conversions.cmake was updated to use add_custom_target() instead of add_custom_command() to create bin and hex files.
There is some kind of bug with xpacks version of ninja-build that breaks on the Windows command line when using add_custom_command().
It has something to do with combining the linker and objcopy commands on a single, funky-looking command line.
Using add_custom_target() instead avoids this attempt by CMake and Ninja to combine it all on one command line.